Move functions from train to dataset
